requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
pyyaml>=6.0
python-multipart>=0.0.6
aiofiles>=23.0
